+++ packages/block-indexer/.env.example
@@ -13,15 +13,17 @@ ASSET_FIXED_NUMBER=6
 DASHBOARD_API_URL=http://localhost:8000/v1
 POSTGRES_DB_URL=postgresql://dashboard:dashboard@54.251.114.18:5432/test_migration
 
-ICON_API_URL=http://54.251.114.18:9080/api/v3/icon
+ICON_API_URL=https://berlin.net.solidwallet.io/api/v3
 # -1 = start from last saved block, 0 = start with head block, > 0 = start with block height
-ICON_BLOCK_HEIGHT=0
-ICON_NETWORK_ID=0x58eb1c
-ICON_FAS_ADDRESS=cx58661408d4ca1298fe38207f0914c8f30171c33b
-ICON_BMC_ADDRESS=cx362bec5623356a90f607d4e693b5d9a898a0dc6f
+ICON_BLOCK_HEIGHT=-1
+ICON_NETWORK_ID=0x7
+ICON_FAS_ADDRESS=cxaccabf74c253e5fb606975a7813d98563443d6d0
+ICON_BMC_ADDRESS=cxf3f7533eead8941a5655a90379caf9d121fb2725
+ICON_HMNY_BMC_ADDRESS=cx17ead9e38e3c67f7e60adffdc2891eab2a932481
 ICON_WPS_BMC=cx9e5c0a749ee94c01febe04702184002a76a84f84
-ICON_WPS_BSH=cx814c72f92d6f23abcf64ca7f2450e012c2199cd3
-ICON_NATIVE_COIN_BSH_ADDRESS=cx489ed02580ce5cab57925317373310205417c2b7
+ICON_WPS_BSH=cxbd91969ee42a7950c3db946246808dfa69e7c298
+ICON_NATIVE_COIN_BSH_ADDRESS=cxe3fbc6d12fc842935ad68991b5247ea821584002
+ICON_NATIVE_COIN_HMNY_BSH_ADDRESS=cxedde5ba834fddcea4ae488eb0e588abb2b7831a9
 ICON_INDEX_INTERVAL=500
 
 MOONBEAM_API_URL=http://54.251.114.18:9933
@@ -34,11 +36,11 @@ MOONBEAM_BMC_MANAGEMENT_ADDRESS=0x3ed62137c5DB927cb137c26455969116BF0c23Cb
 MOONBEAM_BSH_CORE_ADDRESS=0x7d4567B7257cf869B01a47E8cf0EDB3814bDb963
 
 BSC_API_URL=https://data-seed-prebsc-1-s1.binance.org:8545/
-BSC_BLOCK_HEIGHT=18151874
+BSC_BLOCK_HEIGHT=-1
 BSC_NETWORK_ID=0x61
-BSC_BMC_ADDRESS=0x45a0D0cda9e9Fb8e745B91104ca6444DC151D5A7
-BSC_BMC_MANAGEMENT_ADDRESS=0x73005E7919cfE973e0A153Ea09cd6822D0242A42
-BSC_BSH_CORE_ADDRESS=0x5E3f855D3289c52675dFA17a50BdD12E87DE0cf4
+BSC_BMC_ADDRESS=0x121A1AAd623AF68162B1bD84c44234Bc3a3562a9
+BSC_BMC_MANAGEMENT_ADDRESS=0x5C7D42A211b699aa0137Ef589C1f2Beb9F1a5eb8
+BSC_BSH_CORE_ADDRESS=0x84Fbf3F2a497C5be1De18a9d65EeEC297eB7Db41
 
 HARMONY_API_URL=https://rpc.s0.b.hmny.io
 HARMONY_BLOCK_HEIGHT=23702796
