+++ packages/block-indexer/pm2.bsc.json
@@ -4,7 +4,7 @@
       "name": "bsc-indexer",
       "script": "./index.js",
       "args": "bsc",
-      "cwd": "/home/ubuntu/deploy/btp-dashboard/packages/block-indexer",
+      "cwd": "./",
       "log_date_format": "YYYY-MM-DD HH:mm Z",
       "env": {
         "NODE_ENV": "production",
