+++ packages/block-indexer/pm2.icon.json
@@ -4,7 +4,7 @@
       "name": "icon-indexer",
       "script": "./index.js",
       "args": "icon",
-      "cwd": "/home/ubuntu/deploy/btp-dashboard/packages/block-indexer",
+      "cwd": "./",
       "log_date_format": "YYYY-MM-DD HH:mm Z",
       "env": {
         "NODE_ENV": "production",
