+++ packages/block-indexer/src/app/block-indexer.js
@@ -10,7 +10,7 @@ const moonbeamBshAbi = require('../modules/web3-indexer/abi/moonbeam/abi.bsh_cor
 const bscBshAbi = require('../modules/web3-indexer/abi/bsc/BSHPeriphery.json');
 const { getMoonbeamEventMap, getBscEventMap } = require('../modules/common/events');
 const { getMoonbeamActionMap, getBscActionMap } = require('../modules/common/actions');
-const { initBMCAddressMap, initBSHAddressMap } = require('../modules/common/addresses');
+const { setBMCAddressMap, setBSHAddressMap } = require('../modules/common/addresses');
 const logger = createLogger();
 
 async function start(indexerName) {
@@ -20,8 +20,8 @@ async function start(indexerName) {
   const name = indexerName.toUpperCase();
   switch (name) {
     case 'ICON': {
-      initBMCAddressMap();
-      initBSHAddressMap();
+      setBMCAddressMap();
+      setBSHAddressMap();
       iconIndexer.start();
       break;
     }
