--- packages/block-indexer/src/modules/bsc-indexer/abi/BSHPeriphery.json
@@ -1,332 +1,0 @@
-[
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "internalType": "address",
-        "name": "_from",
-        "type": "address"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "_sn",
-        "type": "uint256"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "_code",
-        "type": "uint256"
-      },
-      {
-        "indexed": false,
-        "internalType": "string",
-        "name": "_response",
-        "type": "string"
-      }
-    ],
-    "name": "TransferEnd",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "internalType": "address",
-        "name": "_from",
-        "type": "address"
-      },
-      {
-        "indexed": false,
-        "internalType": "string",
-        "name": "_to",
-        "type": "string"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "_sn",
-        "type": "uint256"
-      },
-      {
-        "components": [
-          {
-            "internalType": "string",
-            "name": "coinName",
-            "type": "string"
-          },
-          {
-            "internalType": "uint256",
-            "name": "value",
-            "type": "uint256"
-          },
-          {
-            "internalType": "uint256",
-            "name": "fee",
-            "type": "uint256"
-          }
-        ],
-        "indexed": false,
-        "internalType": "struct Types.AssetTransferDetail[]",
-        "name": "_assetDetails",
-        "type": "tuple[]"
-      }
-    ],
-    "name": "TransferStart",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": false,
-        "internalType": "string",
-        "name": "_from",
-        "type": "string"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "_sn",
-        "type": "uint256"
-      }
-    ],
-    "name": "UnknownResponse",
-    "type": "event"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "uint256",
-        "name": "",
-        "type": "uint256"
-      }
-    ],
-    "name": "requests",
-    "outputs": [
-      {
-        "internalType": "string",
-        "name": "from",
-        "type": "string"
-      },
-      {
-        "internalType": "string",
-        "name": "to",
-        "type": "string"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function",
-    "constant": true
-  },
-  {
-    "inputs": [],
-    "name": "serviceName",
-    "outputs": [
-      {
-        "internalType": "string",
-        "name": "",
-        "type": "string"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function",
-    "constant": true
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "_bmc",
-        "type": "address"
-      },
-      {
-        "internalType": "address",
-        "name": "_bshCore",
-        "type": "address"
-      },
-      {
-        "internalType": "string",
-        "name": "_serviceName",
-        "type": "string"
-      }
-    ],
-    "name": "initialize",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "hasPendingRequest",
-    "outputs": [
-      {
-        "internalType": "bool",
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function",
-    "constant": true
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "_from",
-        "type": "address"
-      },
-      {
-        "internalType": "string",
-        "name": "_to",
-        "type": "string"
-      },
-      {
-        "internalType": "string[]",
-        "name": "_coinNames",
-        "type": "string[]"
-      },
-      {
-        "internalType": "uint256[]",
-        "name": "_values",
-        "type": "uint256[]"
-      },
-      {
-        "internalType": "uint256[]",
-        "name": "_fees",
-        "type": "uint256[]"
-      }
-    ],
-    "name": "sendServiceMessage",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "string",
-        "name": "_from",
-        "type": "string"
-      },
-      {
-        "internalType": "string",
-        "name": "_svc",
-        "type": "string"
-      },
-      {
-        "internalType": "uint256",
-        "name": "_sn",
-        "type": "uint256"
-      },
-      {
-        "internalType": "bytes",
-        "name": "_msg",
-        "type": "bytes"
-      }
-    ],
-    "name": "handleBTPMessage",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "string",
-        "name": "",
-        "type": "string"
-      },
-      {
-        "internalType": "string",
-        "name": "_svc",
-        "type": "string"
-      },
-      {
-        "internalType": "uint256",
-        "name": "_sn",
-        "type": "uint256"
-      },
-      {
-        "internalType": "uint256",
-        "name": "_code",
-        "type": "uint256"
-      },
-      {
-        "internalType": "string",
-        "name": "_msg",
-        "type": "string"
-      }
-    ],
-    "name": "handleBTPError",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "string",
-        "name": "_to",
-        "type": "string"
-      },
-      {
-        "components": [
-          {
-            "internalType": "string",
-            "name": "coinName",
-            "type": "string"
-          },
-          {
-            "internalType": "uint256",
-            "name": "value",
-            "type": "uint256"
-          }
-        ],
-        "internalType": "struct Types.Asset[]",
-        "name": "_assets",
-        "type": "tuple[]"
-      }
-    ],
-    "name": "handleRequestService",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "string",
-        "name": "_fa",
-        "type": "string"
-      },
-      {
-        "internalType": "string",
-        "name": "_svc",
-        "type": "string"
-      }
-    ],
-    "name": "handleFeeGathering",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "string",
-        "name": "_to",
-        "type": "string"
-      }
-    ],
-    "name": "checkParseAddress",
-    "outputs": [],
-    "stateMutability": "pure",
-    "type": "function",
-    "constant": true
-  }
-]
