--- packages/block-indexer/src/modules/bsc-indexer/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const indexer = require('./block-indexer');
-
-module.exports = {
-  ...indexer
-};
