+++ packages/block-indexer/src/modules/mint-burn/icon.js
@@ -1,9 +1,8 @@
-/* eslint-disable curly */
-/* eslint-disable yoda */
 'use strict';
 
 const { IconConverter } = require('icon-sdk-js').default;
 const { createLogger, ICX_LOOP_UNIT, MINT_EVENT, BURN_EVENT, ICON_ZERO_ADDRESS } = require('../../common');
+const { getBMCAddressesMap } = require('../common/addresses');
 const { getTokenName } = require('../tokens/model');
 const { getTotalTokenAmount, saveToken } = require('./repository');
 
@@ -34,9 +33,9 @@ async function getMintBurnEvent(event, transaction) {
 // Mint: Alice on ICON got a DEV from Moonbeam (a DEV minted to Alice).
 // Burn: Alice on ICON send a DEV (which she got from Bob earlier) back to Bob on Moonbeam (a DEV burned from Alice).
 async function handleMintBurnEvents(txResult, transaction) {
-  if (process.env.ICON_BMC_ADDRESS !== transaction.to || 0 === txResult.eventLogs.length ||
-    TRANSFER_PROTOTYPE !== txResult.eventLogs[0].indexed[0])
-    return false;
+  const bmcAddressesMap = getBMCAddressesMap();
+  if (!bmcAddressesMap.has(transaction.to) || txResult.eventLogs.length === 0 ||
+    TRANSFER_PROTOTYPE !== txResult.eventLogs[0].indexed[0]) { return false; }
 
   try {
     const mintBurn = await getMintBurnEvent(txResult.eventLogs[0], transaction);
