+++ packages/dashboard-api/.env.example
@@ -16,7 +16,7 @@ ICON_NETWORK_ID=0x7
 ICON_NAME=ICON
 ICON_NATIVE_TOKEN=ICX
 ICON_FAS_ADDRESS=cxaccabf74c253e5fb606975a7813d98563443d6d0
-ICON_BMC_ADDRESS=cxa2cc386e9db2a72ea6724cbfd12f936a90ba63d2
+ICON_BMC_ADDRESS=cxf3f7533eead8941a5655a90379caf9d121fb2725
 ICON_NETWORK_PATH_LOGO=./image/logo/icon-icx-logo.png
 ICON_NETWORK_URL=https://icon.foundation
