+++ packages/dashboard-api/pm2.json
@@ -3,7 +3,7 @@
     {
       "name": "dashboard-api",
       "script": "./index.js",
-      "cwd": "/home/ubuntu/deploy/btp-dashboard/packages/dashboard-api",
+      "cwd": "./",
       "log_date_format": "YYYY-MM-DD HH:mm Z",
       "env": {
         "NODE_ENV": "production"
